project_sample/backend_django/apps/sample/serializers.py
from rest_framework.serializers import ModelSerializer

from .models import People


class PeopleSerializer(ModelSerializer):
    class Meta:
        model = People
        fields = [
            'id',
            'first_name',
            'last_name',
        ]
