project_sample/backend_django/requirements/base.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
mysqlclient==2.0.3
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
